+++ react/ProductSummaryCustom.tsx
@@ -35,6 +35,7 @@ function ProductSummaryCustom({
   children,
   href,
   priceBehavior = 'default',
+  position,
   classes,
 }: PropsWithChildren<Props>) {
   const {
@@ -58,14 +59,14 @@ function ProductSummaryCustom({
   const onView = useCallback(() => {
     productListDispatch?.({
       type: 'SEND_IMPRESSION',
-      args: { product },
+      args: { product, position },
     })
 
     dispatch({
       type: 'SET_IN_VIEW',
       args: { inView: true },
     })
-  }, [dispatch, productListDispatch, product])
+  }, [dispatch, productListDispatch, product, position])
 
   useOnView({
     ref: inViewRef,
@@ -186,6 +187,20 @@ interface Props {
    * @default "default"
    */
   priceBehavior?: 'async' | 'default'
+  /**
+   * Name of the list the Product Summary is in. Should be set by a Shelf or the Search Result gallery.
+   */
+  listName?: string
+  /**
+   * Whether the listName should be sent to product pages via querystring
+   * @default true
+   */
+  trackListName?: boolean
+  /**
+   * The position of the Product Summary in a list of Product Summaries. Used by the
+   * ProductImpressions event.
+   */
+  position?: number
   classes?: CssHandlesTypes.CustomClasses<typeof CSS_HANDLES>
 }
 
@@ -194,12 +209,16 @@ function ProductSummaryWrapper({
   actionOnClick,
   href,
   priceBehavior = 'default',
+  trackListName = true,
+  listName,
+  position,
   classes,
   children,
 }: PropsWithChildren<Props>) {
   return (
     <ProductSummaryProvider
       product={product}
+      listName={trackListName ? listName : undefined}
       isPriceLoading={priceBehavior === 'async'}
     >
       <ProductSummaryCustom
@@ -207,6 +226,7 @@ function ProductSummaryWrapper({
         href={href}
         actionOnClick={actionOnClick}
         priceBehavior={priceBehavior}
+        position={position}
         classes={classes}
       >
         {children}