+++ react/ProductSummarySKUSelector.tsx
@@ -7,7 +7,7 @@ import type { ProductSummaryTypes } from 'vtex.product-summary-context'
 import type { ResponsiveValuesTypes } from 'vtex.responsive-values'
 import type { ProductTypes } from 'vtex.product-context'
 
-import { getFirstAvailableSeller } from './modules/seller'
+import { getDefaultSeller } from './modules/seller'
 
 const { useProductSummary, useProductSummaryDispatch } = ProductSummaryContext
 
@@ -189,9 +189,9 @@ function ProductSummarySKUSelector(props: Props) {
     const sku = {
       ...selectedItem,
       image: selectedItem.images[0],
-      seller:
-        getFirstAvailableSeller(selectedItem.sellers) ??
-        selectedItem.sellers[0],
+      seller: getDefaultSeller(
+        selectedItem.sellers
+      ) as ProductSummaryTypes.Seller,
     }
 
     const newProduct = {