+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Seller to be the default seller instead of the first available seller.
+- `actionOnClick` not being passed to `ProductSummary` slot.
+- Product summary position now starts at 1.
+
+### Added
+- Position and list name to `productClick` GTM event.
+- `ProductSummaryCustom` now accepts the `listName` and `position` props. This is intended to be used by Shelves and Search Result galleries.
+- `trackListName` prop to `ProductSummaryCustom`. It controls whether it should pass the `listName` prop to the PDP url or not.
 
 ## [2.72.0] - 2021-05-03
 ### Added